requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --output-file=requirements.txt
-e .
alembic==1.16.5
cfgv==3.4.0
cloudpickle==3.1.1
colorama==0.4.6 ; sys_platform == 'win32'
colorlog==6.9.0
coverage==7.8.0
distlib==0.3.9
dpkt==1.9.8
duckdb==1.2.1
filelock==3.18.0
greenlet==3.2.4 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
identify==2.6.9
iniconfig==2.1.0
joblib==1.5.2
llvmlite==0.44.0
lxml==5.3.1
mako==1.3.10
markupsafe==3.0.2
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.5
nvidia-nccl-cu12==2.27.7 ; platform_machine != 'aarch64' and sys_platform == 'linux'
optuna==4.5.0
packaging==24.2
pandas==2.3.2
pcap-filter==0.1.0
platformdirs==4.3.7
pluggy==1.5.0
polars-lts-cpu==1.28.1
pre-commit==4.2.0
pyarrow==19.0.1
pyright==1.1.398
pytest==8.3.5
pytest-cov==6.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
ruff==0.11.3
scikit-learn==1.7.1
scipy==1.16.1
shap==0.48.0
six==1.17.0
slicer==0.0.8
sqlalchemy==2.0.43
threadpoolctl==3.6.0
tqdm==4.67.1
typing-extensions==4.13.1
tzdata==2025.2
virtualenv==20.30.0
xgboost==3.0.5
